requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
pynput>=1.7.6
pyyaml>=6.0.1
numpy>=1.24.0 